#12 플레이어스토어에 player와 nowTurn getter 로직 설정 및 type 파일로 별도 정의
